introduce HeaderContent
